adjust rate from int to double, and reduce rate to 0.2 per wheel

+++ imgviewer/imgviewer.h
@@ -44,7 +44,7 @@ private:
     Base::NAutoPtr<NFont> font_;
     Base::NAutoPtr<NImageDraw> image_;
 
-    int rate_;
+    double rate_;
 
     NFileFinder fileFinder_;
